More fixes

+++ examples/semi_async_pipeline_coordinator.cpp
@@ -223,7 +223,7 @@ int main() {
     std::vector<Tensor<float>> micro_batch_labels =
         batch_labels.split(mnist_constants::NUM_MICROBATCHES);
 
-    for (int i = 0; i < micro_batches.size(); ++i) {
+    for (size_t i = 0; i < micro_batches.size(); ++i) {
       coordinator.forward(micro_batches[i], i);
     }
